Modified ELUCData by consolidating yucky classes into single clean class
